# src/main/resources/sandboxOpenApiSchema.yaml
openapi: 3.0.0
info:
  title: API for budget application
  description: OpenAPI schema of budget application
  version: 0.1.9
tags:
  - name: hello world sample
paths:
  /hello:
    get:
      summary: Returns a 'hello world' string.
      tags:
        - hello world sample
      description: Returns a 'hello world' string.
      operationId: hello
      responses:
        '200':
          description: A list of user's wallets
          content:
            application/json:
              schema:
                type: string
                example: hello world
  /wallet:
    get:
      summary: Returns  of user's wallets.
      tags:
        - wallet
      description: Returns a list of user's wallets.
      operationId: viewWallets
      responses:
        '200':
          description: A list of user's wallets
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/wallet'
  /api/authentication/login:
    post:
      tags:
        - User Management
      summary: LogIn
      description: As a User I want to to have a login functionality so that I can get an access to my User account
      operationId: logIn
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LogIn'
        required: true
      responses:
        200:
          description: ok
        401:
          description: Invalid Email or Password
        422:
          description: Wrong Email or Password format


components:
    schemas:
      wallet:
        type: object
        properties:
          id:
            type: integer
            format: int64
          balance:
            type: integer
            format: int64
          currency:
            type: string
          is_default:
            type: boolean
      LogIn:
        type: object
        properties:
          email:
            type: string
            maxLength: 254
            pattern: ^[a-z0-9]+[\.]{0,1}[a-z0-9]+@[a-z0-9]+\.[a-z]{2,4}$
          password:
            type: string
            minLength: 8
            maxLength: 100
            pattern: ^[a-zA-Z0-9!@#\$%\^\&*\)\(/+=._-]
    securitySchemes:
      bearerAuth:
        type: http
        description: JWT Authorization header using the Bearer scheme
        scheme: bearer
        bearerFormat: JWT


security:
   - bearerAuth: [ ]